pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "codefix-env"
version = "1.0.0"
description = "Code debugging RL environment for evaluating AI debugging agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.11.0",
    "requests>=2.32.0",
    "openai>=1.75.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["env*", "graders*", "tasks*", "utils*", "server*"]
